[ADD / UPDATE] Graphic API working
TurtleNormal doesnt have headless mode,
so it cannot save the drawing as a SVG without a render

Linux cursor hover crashing due to rust version or stuff..

+++ src/action.rs
@@ -1,16 +1,16 @@
 #[derive(Clone)]
 pub enum Do {
-    Forward(usize),
-    Backward(usize),
-    Left(usize),
-    Right(usize),
+    Forward(f64),
+    Backward(f64),
+    Left(f64),
+    Right(f64),
     PenUp,
     PenDown,
     TurnRandom,
     ColorRandom,
     Save,
     Restore,
-    IncreaseSize(usize),
-    DecreaseSize(usize),
-    LineSize(usize)
+    IncreaseSize(f64),
+    DecreaseSize(f64),
+    LineSize(f64)
 }
